License under GPLv3
I followed the [guide from the GNU website][6] in order to license it.

JavaFX is [licensed under GPL v2 with classpath exception][1], and I'm
only using it as a library so I don't need to follow this license.
NetBeans is [licensed under Apache v2][2] ([compatible with the
GPL][5]), I have to assume that this applies to the generated project
source code. The Scene Builder is [licensed under BSD 3-Clause][3] which
is [compatible with the GPL][4].

[1]: https://web.archive.org/web/20220310233018/https://gluonhq.com/products/javafx/
[2]: https://github.com/apache/netbeans/blob/13/LICENSE
[3]: https://web.archive.org/web/20220303112352/https://gluonhq.com/products/scene-builder/
[4]: https://www.gnu.org/licenses/license-list.html#ModifiedBSD
[5]: https://www.gnu.org/licenses/license-list.html#apache2
[6]: https://www.gnu.org/licenses/gpl-howto.html